amproxy/requirements.txt
python-dateutil==2.8.0
flask==1.0.2
requests==2.22.0